+++ tests/helpers.js
@@ -73,3 +73,54 @@ export const clearCookie = (cName) => {
 };
 
 export const isSafariBrowser = () => navigator.vendor === 'Apple Computer, Inc.';
+
+export const PANEL_ID = 'panel';
+export const CLICKABLE_NAME = 'clickable';
+export const SELECTORS_DELIMITER = ',';
+
+/**
+ * Generates a CSS selector string based on the order of clicked elements.
+ *
+ * @param {number[]} clickOrder - An array of numbers representing the order in which elements were clicked.
+ * @returns {string} A string of CSS selectors corresponding to the click order.
+ */
+export const createSelectorsString = (clickOrder) => {
+    const selectors = clickOrder.map((elemNum) => `#${PANEL_ID} > #${CLICKABLE_NAME}${elemNum}`);
+    return selectors.join(SELECTORS_DELIMITER);
+};
+
+/**
+ * Creates a clickable checkbox element with a unique ID and an onClick assertion.
+ *
+ * @param {number} elementNum - The number for this element.
+ * @param {string} [text=''] - Optional text content for the checkbox.
+ * @returns {HTMLInputElement} The created checkbox element.
+ */
+export const createClickable = (elementNum, text = '') => {
+    const clickableId = `${CLICKABLE_NAME}${elementNum}`;
+    const checkbox = document.createElement('input');
+    checkbox.type = 'checkbox';
+    checkbox.id = clickableId;
+    checkbox.textContent = text;
+    checkbox.onclick = (e) => {
+        e.currentTarget.setAttribute('clicked', true);
+        window.clickOrder.push(elementNum);
+    };
+    return checkbox;
+};
+/**
+ * Creates a panel div element with the specified ID and appends it to the body.
+ *
+ * @returns {HTMLDivElement} The created panel element.
+ */
+export const createPanel = () => {
+    const panel = document.createElement('div');
+    panel.id = PANEL_ID;
+    document.body.appendChild(panel);
+    return panel;
+};
+
+/**
+ * Removes the panel element from the document.
+ */
+export const removePanel = () => document.getElementById('panel').remove();