+++ tests/scriptlets/index.test.js
@@ -47,6 +47,7 @@ import './trusted-replace-xhr-response.test';
 import './xml-prune.test';
 import './m3u-prune.test';
 import './trusted-click-element.test';
+import './trusted-click-element.spec';
 import './trusted-set-cookie.test';
 import './trusted-replace-fetch-response.test';
 import './trusted-set-local-storage-item.test';